Split Canvas in multiple components
